added p-values
functions in tda_and_scatter.R now calculate the p-value for the two variables on the scatterplot. Also added the contour_difference.R code, which is just a playground of figuring out ggplot2 and how to make marginal plots. to do: final touches on density_comparison function to plot contours instead of points, maybe?